fix(click-away): fix click away detection inside shadow DOM
